Tema07 - Proyecto Django/Clase16/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv